excel_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openpyxl==3.1.2
python-multipart==0.0.6
PyJWT==2.8.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0 